refactor(oidc-provider): move sts_session_name into exchange::aws
`sts_session_name` encodes an AWS-STS `RoleSessionName` charset constraint,
but lived at module scope in `backend_auth.rs` — which is explicitly
multi-backend (the `match backend_type` already stubs non-`s3` backends).
Relocate it (and its unit test) next to `AwsExchange::session_name` so the
RoleSessionName rule stays single-sourced in the AWS module and out of the
way of future GCP/Azure backends. Pure move: `pub(crate)` so `resolve_aws`
can still call it; the end-to-end `session_name_reaches_sts_request` test
stays with the middleware it exercises.
