+++ crates/oidc-provider/src/backend_auth.rs
@@ -15,7 +15,7 @@ use multistore::types::BucketConfig;
 use std::borrow::Cow;
 use std::collections::HashMap;
 
-use crate::exchange::aws::AwsExchange;
+use crate::exchange::aws::{sts_session_name, AwsExchange};
 use crate::{HttpExchange, OidcCredentialProvider};
 
 /// AWS OIDC backend auth — exchanges a self-signed JWT for temporary
@@ -157,34 +157,6 @@ impl<H: HttpExchange> Middleware for MaybeOidcAuth<H> {
     }
 }
 
-/// Map an OIDC subject to a valid AWS `RoleSessionName`.
-///
-/// `RoleSessionName` must match `[\w+=,.@-]{2,64}`; the proxy's subject
-/// (`scv1:conn:{id}`) contains `:`, which STS rejects. Replace any
-/// disallowed character with `_` and clamp to 64 chars. This is only a
-/// CloudTrail attribution label, not a security boundary (the trust policy
-/// gates on the JWT `sub`/`aud`, not the session name), so the rare
-/// truncation collision is cosmetic. Falls back to `s3-proxy` if sanitizing
-/// leaves fewer than the 2 chars STS requires.
-fn sts_session_name(subject: &str) -> String {
-    let name: String = subject
-        .chars()
-        .map(|c| {
-            if c.is_ascii_alphanumeric() || "+=,.@-_".contains(c) {
-                c
-            } else {
-                '_'
-            }
-        })
-        .take(64)
-        .collect();
-    if name.len() < 2 {
-        "s3-proxy".to_string()
-    } else {
-        name
-    }
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
@@ -281,26 +253,6 @@ mod tests {
             allowed_roles: vec![],
             backend_options: opts,
         }
-    }
-
-    #[test]
-    fn session_name_sanitizes_subject() {
-        // Colons (and any other disallowed char) become `_`; allowed chars
-        // (alnum, `-`, `.`, `_`, `@`, `+`, `=`, `,`) pass through.
-        assert_eq!(sts_session_name("scv1:conn:abc-123"), "scv1_conn_abc-123");
-        // Already-valid default is untouched.
-        assert_eq!(sts_session_name("s3-proxy"), "s3-proxy");
-        // Clamped to AWS's 64-char ceiling.
-        assert_eq!(sts_session_name(&"a".repeat(100)).len(), 64);
-        // Degenerate subject falls back rather than emit an invalid (<2 char) name.
-        assert_eq!(sts_session_name(":"), "s3-proxy");
-        // The minted name is always within AWS's [2,64] RoleSessionName bounds
-        // and contains only permitted characters.
-        let n = sts_session_name("scv1:conn:00000000-0000-0000-0000-000000000000");
-        assert!((2..=64).contains(&n.len()));
-        assert!(n
-            .chars()
-            .all(|c| c.is_ascii_alphanumeric() || "+=,.@-_".contains(c)));
     }
 
     #[tokio::test]
